old KV implementation marked as deprecated
