+++ models/classes/execution/class.KeyValueService.php
@@ -18,7 +18,7 @@
  *
  */
 
-use oat\oatbox\Configurable;
+use oat\taoDelivery\model\execution\KeyValueService;
 use oat\taoDelivery\models\classes\execution\DeliveryExecution;
 use oat\taoDelivery\model\execution\DeliveryExecution as InterfaceDeliveryExecution;
 /**
@@ -27,31 +27,10 @@ use oat\taoDelivery\model\execution\DeliveryExecution as InterfaceDeliveryExecut
  * @access public
  * @author Joel Bout, <joel@taotesting.com>
  * @package taoDelivery
+ * @deprecated  please use oat\taoDelivery\model\execution\KeyValueService
  */
-class taoDelivery_models_classes_execution_KeyValueService extends Configurable
-    implements taoDelivery_models_classes_execution_Service
+class taoDelivery_models_classes_execution_KeyValueService extends KeyValueService
 {
-    const OPTION_PERSISTENCE = 'persistence';
-
-    const DELIVERY_EXECUTION_PREFIX = 'kve_de_';
-
-    const USER_EXECUTIONS_PREFIX = 'kve_ue_';
-
-    /**
-     * @var common_persistence_KeyValuePersistence
-     */
-    private $persistence;
-
-    protected function getPersistence()
-    {
-        if (is_null($this->persistence)) {
-            $persistenceOption = $this->getOption(self::OPTION_PERSISTENCE);
-            $this->persistence = (is_object($persistenceOption))
-                ? $persistenceOption
-                : common_persistence_KeyValuePersistence::getPersistence($persistenceOption);
-        }
-        return $this->persistence;
-    }
 
     public function getUserExecutions(core_kernel_classes_Resource $compiled, $userUri)
     {
@@ -64,21 +43,6 @@ class taoDelivery_models_classes_execution_KeyValueService extends Configurable
                 $returnValue[] = $de;
             }
         }
-        return $returnValue;
-    }
-
-    public function getDeliveryExecutionsByStatus($userUri, $status) {
-        $returnValue = array();
-        $data = $this->getPersistence()->get(self::USER_EXECUTIONS_PREFIX.$userUri.$status);
-        $keys = $data !== false ? json_decode($data) : array();
-        if (is_array($keys)) {
-            foreach ($keys as $key) {
-                $returnValue[$key] = $this->getDeliveryExecution($key);
-            }
-        } else {
-            common_Logger::w('Non array "'.gettype($keys).'" received as active Delivery Keys for user '.$userUri);
-        }
-
         return $returnValue;
     }
 
@@ -97,51 +61,5 @@ class taoDelivery_models_classes_execution_KeyValueService extends Configurable
         $this->updateDeliveryExecutionStatus($deliveryExecution, null, InterfaceDeliveryExecution::STATE_ACTIVE);
 
         return $deliveryExecution;
-    }
-
-    public function getDeliveryExecution($identifier) {
-        $deImplementation = new \taoDelivery_models_classes_execution_KVDeliveryExecution($this->getPersistence(), $identifier);
-        return new DeliveryExecution($deImplementation);
-    }
-
-    /**
-     * Update the collection of deliveries
-     *
-     * @param DeliveryExecution $deliveryExecution
-     * @param string $old
-     * @param string $new
-     */
-    public function updateDeliveryExecutionStatus(DeliveryExecution $deliveryExecution, $old, $new) {
-
-        $userId = $deliveryExecution->getUserIdentifier();
-        if ($old != null) {
-            $oldReferences = $this->getDeliveryExecutionsByStatus($userId, $old);
-            foreach (array_keys($oldReferences) as $key) {
-                if ($oldReferences[$key]->getIdentifier() == $deliveryExecution->getIdentifier()) {
-                    unset($oldReferences[$key]);
-                }
-            }
-            $this->setDeliveryExecutions($userId, $old, $oldReferences);
-        }
-
-        $newReferences = $this->getDeliveryExecutionsByStatus($userId, $new);
-        $newReferences[] = $deliveryExecution;
-        $this->setDeliveryExecutions($userId, $new, $newReferences);
-
-    }
-
-    public function getData($deliveryExecutionId) {
-        $dataString = $this->getPersistence()->get($deliveryExecutionId);
-        $data = json_decode($dataString, true);
-        return $data;
-    }
-
-    private function setDeliveryExecutions($userUri, $status, $executions)
-    {
-        $keys = array();
-        foreach ($executions as $execution) {
-            $keys[] = $execution->getIdentifier();
-        }
-        return $this->getPersistence()->set(self::USER_EXECUTIONS_PREFIX.$userUri.$status, json_encode($keys));
     }
 }
